+++ CHANGELOG.md
@@ -7,7 +7,12 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `EnvelopeDocumentsEndpoint.merge`
+- Add `EnvelopeTemplatesEndpoint.clone`
 - Add `AccountMeEndpoint.update`
+- Add `EnvelopeProperties` resource and `Envelope.properties`
+- Add `EnvelopeTemplateTag.assignment` + `EnvelopeTemplateTag.EnvelopeRecipient` can be null
+- Add `EnvelopeTag.assignment` + `EnvelopeTag.EnvelopeRecipient` can be null
+- Add `EnvelopeTemplate.useDefaultTemplateSettings`
 - Add `AccountSettings.bankIdSign`
 
 ## [1.4.0] - 2021-09-17